refactor: extract WorkspaceState timeline into an extension file
Move the timeline subsystem (message load + pagination windowing, reply
drafts, reactions, delete, send, read-marking, message-cache pruning, the
initial-load gate, and the timeline subscription listener) into
WorkspaceState+Timeline.swift. Pure code movement, no behavior change.
Build green; all 208 tests pass.
