+++ whitenoise-mac/Core/WorkspaceState.swift
@@ -1246,178 +1246,6 @@ final class WorkspaceState {
         }
     }
 
-    func loadMessages(groupIdHex: String) async {
-        guard let client, let activeAccount else {
-            finishTimelineInitialLoad(groupIdHex: groupIdHex)
-            return
-        }
-        if timelineTaskGroupId == groupIdHex, messagesByChat[groupIdHex] != nil {
-            finishTimelineInitialLoad(groupIdHex: groupIdHex)
-            return
-        }
-        stopTimelineListener()
-        guard selectedChat?.id == groupIdHex else {
-            finishTimelineInitialLoad(groupIdHex: groupIdHex)
-            return
-        }
-        beginTimelineInitialLoadIfNeeded(groupIdHex: groupIdHex)
-        defer { finishTimelineInitialLoad(groupIdHex: groupIdHex) }
-        do {
-            let accountRef = activeAccount.accountRef
-            if let row = try await runOffMain({
-                try client.initializeChatReadState(accountRef: accountRef, groupIdHex: groupIdHex)
-            }) {
-                // `initializeChatReadState` may race a live chat-list delta. Do not let an
-                // older read-state row roll back a newer preview/timestamp already applied
-                // by the subscription listener while the FFI call was in flight.
-                await applyChatRow(
-                    row,
-                    account: activeAccount,
-                    skippingStaleRow: true,
-                    shouldEnrich: false
-                )
-            }
-
-            let subscription = try await client.subscribeTimelineMessages(
-                accountRef: activeAccount.accountRef,
-                groupIdHex: groupIdHex,
-                limit: Self.timelinePageLimit
-            )
-            guard activeAccountId == activeAccount.id, selectedChat?.id == groupIdHex else { return }
-
-            let page =
-                subscription.snapshot()
-                ?? TimelinePageFfi(
-                    messages: [],
-                    hasMoreBefore: false,
-                    hasMoreAfter: false
-                )
-            await applyTimelineWindow(page, groupIdHex: groupIdHex, account: activeAccount, client: client)
-            // Start the listener first (it tears down any prior listener, which would clear
-            // these), then record the subscription so scroll-back pagination can reach it.
-            // `startTimelineListener` can bail without starting a task (e.g. selection changed
-            // while we awaited above); only record the subscription when it actually started,
-            // otherwise we leak a live handle with no `next()` loop draining it.
-            startTimelineListener(groupIdHex: groupIdHex, account: activeAccount, subscription: subscription)
-            guard timelineTaskGroupId == groupIdHex else { return }
-            activeTimelineSubscription = subscription
-            activeTimelineGroupId = groupIdHex
-        } catch {
-            lastError = error.localizedDescription
-        }
-    }
-
-    func loadOlderMessages(groupIdHex: String) async {
-        guard let client, let activeAccount else { return }
-        guard selectedChat?.id == groupIdHex, activeTimelineGroupId == groupIdHex else { return }
-        guard let subscription = activeTimelineSubscription else { return }
-        guard var paging = timelinePagingByChat[groupIdHex],
-            paging.hasMoreBefore,
-            !paging.isLoadingBefore
-        else { return }
-
-        paging.isLoadingBefore = true
-        timelinePagingByChat[groupIdHex] = paging
-        defer {
-            if var currentPaging = timelinePagingByChat[groupIdHex] {
-                currentPaging.isLoadingBefore = false
-                timelinePagingByChat[groupIdHex] = currentPaging
-            }
-        }
-
-        do {
-            // The subscription owns the materialized window; `paginateBackwards` extends it
-            // toward older history off the main thread and returns the new authoritative
-            // window (already sorted, deduped, capped, with correct has-more flags).
-            let page = try await subscription.paginateBackwards(count: Self.timelinePageLimit)
-            guard activeAccountId == activeAccount.id, selectedChat?.id == groupIdHex else { return }
-            await applyTimelineWindow(page, groupIdHex: groupIdHex, account: activeAccount, client: client)
-        } catch {
-            lastError = error.localizedDescription
-        }
-    }
-
-    func loadNewerMessages(groupIdHex: String) async {
-        guard let client, let activeAccount else { return }
-        guard selectedChat?.id == groupIdHex, activeTimelineGroupId == groupIdHex else { return }
-        guard let subscription = activeTimelineSubscription else { return }
-        guard var paging = timelinePagingByChat[groupIdHex],
-            paging.hasMoreAfter,
-            !paging.isLoadingAfter
-        else { return }
-
-        paging.isLoadingAfter = true
-        timelinePagingByChat[groupIdHex] = paging
-        defer {
-            if var currentPaging = timelinePagingByChat[groupIdHex] {
-                currentPaging.isLoadingAfter = false
-                timelinePagingByChat[groupIdHex] = currentPaging
-            }
-        }
-
-        do {
-            let page = try await subscription.paginateForwards(count: Self.timelinePageLimit)
-            guard activeAccountId == activeAccount.id, selectedChat?.id == groupIdHex else { return }
-            await applyTimelineWindow(page, groupIdHex: groupIdHex, account: activeAccount, client: client)
-        } catch {
-            lastError = error.localizedDescription
-        }
-    }
-
-    /// Render an authoritative timeline window from the subscription (initial snapshot,
-    /// pagination result, or live update). The window is already ordered/deduped/capped by
-    /// the runtime, so we map + resolve senders and replace the transcript wholesale.
-    func applyTimelineWindow(
-        _ page: TimelinePageFfi,
-        groupIdHex: String,
-        account: AccountItem,
-        client: any MarmotRuntime
-    ) async {
-        guard activeAccountId == account.id, selectedChat?.id == groupIdHex else { return }
-        let senderProfiles = await messageSenderProfiles(
-            from: page.messages,
-            groupIdHex: groupIdHex,
-            activeAccount: account,
-            client: client
-        )
-        guard activeAccountId == account.id, selectedChat?.id == groupIdHex else { return }
-
-        let currentPaging = timelinePagingByChat[groupIdHex]
-        replaceMessages(
-            MessageItem.timeline(
-                from: page,
-                activeAccountIdHex: account.accountIdHex,
-                senderProfiles: senderProfiles
-            ),
-            groupIdHex: groupIdHex,
-            paging: TimelinePagingState(
-                hasMoreBefore: page.hasMoreBefore,
-                hasMoreAfter: page.hasMoreAfter,
-                isLoadingBefore: currentPaging?.isLoadingBefore ?? false,
-                isLoadingAfter: currentPaging?.isLoadingAfter ?? false
-            )
-        )
-        await markLatestVisibleMessageRead(groupIdHex: groupIdHex, account: account, client: client)
-    }
-
-    func startReply(to message: MessageItem) {
-        guard message.supportsChatActions else { return }
-        replyDraftContext = MessageReplyContext(
-            targetMessageId: message.id,
-            senderName: message.senderName,
-            body: message.replyPreviewText
-        )
-    }
-
-    func cancelReply() {
-        replyDraftContext = nil
-    }
-
-    func copyText(of message: MessageItem) {
-        guard message.canCopyText else { return }
-        copyText(message.body)
-    }
-
     func mediaDownloadState(for message: MessageItem, attachment: MessageMediaAttachment) -> MediaDownloadState {
         mediaDownloadStateStore(for: message, attachment: attachment).state
     }
@@ -1490,18 +1318,6 @@ final class WorkspaceState {
         return store
     }
 
-    /// Copies `text` to the system pasteboard.
-    ///
-    /// Every value copied from this app is private-messenger content — decrypted message
-    /// bodies, full conversation transcripts, and Nostr identity keys — so copies default to
-    /// `concealed`. A concealed copy additionally carries the `org.nspasteboard.ConcealedType`
-    /// marker (see `copyToGeneralPasteboard`), which clipboard-history managers honor to avoid
-    /// persisting the value and which discourages Universal Clipboard (Handoff) from syncing it
-    /// to the user's other devices.
-    func copyText(_ text: String, concealed: Bool = true) {
-        copyTextHandler(text, concealed)
-    }
-
     /// The bech32 `npub` form of a hex public key — the canonical, user-facing way to show
     /// a Nostr public key. Falls back to the hex until the account cache has been hydrated
     /// off the main thread.
@@ -1509,68 +1325,6 @@ final class WorkspaceState {
         let trimmed = accountIdHex.trimmingCharacters(in: .whitespacesAndNewlines)
         guard !trimmed.isEmpty else { return trimmed }
         return accounts.first(where: { $0.accountIdHex == trimmed })?.npub ?? trimmed
-    }
-
-    func react(to message: MessageItem, emoji: String) async {
-        guard message.supportsChatActions else { return }
-        guard let client, let activeAccount, let selectedChat else { return }
-        // Reentrancy guard: drop a duplicate of the *same* in-flight reaction
-        // (same target + emoji) while allowing a different emoji on the same message.
-        let reactionKey = "\(selectedChat.id)\u{1F}\(message.id)\u{1F}\(emoji)"
-        guard !inFlightReactionKeys.contains(reactionKey) else { return }
-        inFlightReactionKeys.insert(reactionKey)
-        defer { inFlightReactionKeys.remove(reactionKey) }
-        do {
-            _ = try await client.reactToMessage(
-                accountRef: activeAccount.accountRef,
-                groupIdHex: selectedChat.id,
-                targetMessageId: message.id,
-                emoji: emoji
-            )
-        } catch {
-            lastError = error.localizedDescription
-        }
-    }
-
-    func removeReaction(_ reaction: MessageReaction, from message: MessageItem) async {
-        guard message.supportsChatActions else { return }
-        guard reaction.canRemoveOwnReaction, let reactionMessageId = reaction.ownReactionMessageId else { return }
-        guard let client, let activeAccount, let selectedChat else { return }
-        // Reentrancy guard: the removal deletes the reaction event, so key on its id
-        // (shared namespace with `deleteMessage`) to drop a repeated in-flight removal.
-        guard !inFlightDeleteMessageIds.contains(reactionMessageId) else { return }
-        inFlightDeleteMessageIds.insert(reactionMessageId)
-        defer { inFlightDeleteMessageIds.remove(reactionMessageId) }
-        do {
-            _ = try await client.deleteMessage(
-                accountRef: activeAccount.accountRef,
-                groupIdHex: selectedChat.id,
-                targetMessageId: reactionMessageId
-            )
-        } catch {
-            lastError = error.localizedDescription
-        }
-    }
-
-    func deleteMessage(_ message: MessageItem) async {
-        guard message.canDelete else { return }
-        guard let client, let activeAccount, let selectedChat else { return }
-        // Reentrancy guard: drop a repeated delete of the same in-flight message.
-        guard !inFlightDeleteMessageIds.contains(message.id) else { return }
-        inFlightDeleteMessageIds.insert(message.id)
-        defer { inFlightDeleteMessageIds.remove(message.id) }
-        do {
-            _ = try await client.deleteMessage(
-                accountRef: activeAccount.accountRef,
-                groupIdHex: selectedChat.id,
-                targetMessageId: message.id
-            )
-            if replyDraftContext?.targetMessageId == message.id {
-                replyDraftContext = nil
-            }
-        } catch {
-            lastError = error.localizedDescription
-        }
     }
 
     func groupDetailsSnapshot(
@@ -1841,59 +1595,6 @@ final class WorkspaceState {
         }
     }
 
-    func sendDraft() async {
-        let text = draftText.trimmingCharacters(in: .whitespacesAndNewlines)
-        let mediaAttachments = pendingMediaAttachments
-        // `!isSending` is the reentrancy guard: `isSending` flips synchronously here,
-        // but the model only suspends (and `draftText` is only cleared) at the `await`
-        // below. Without this guard a second invocation delivered before SwiftUI
-        // re-renders the disabled send button (⌘-Return auto-repeat, double events)
-        // would still observe the old `draftText` and re-send the same message.
-        guard let client,
-            let activeAccount,
-            let selectedChat,
-            let draftKey = selectedComposerDraftKey,
-            !text.isEmpty || !mediaAttachments.isEmpty,
-            !isSending
-        else { return }
-        isSending = true
-        defer { isSending = false }
-
-        do {
-            if !mediaAttachments.isEmpty {
-                _ = try await client.uploadMedia(
-                    accountRef: activeAccount.accountRef,
-                    groupIdHex: selectedChat.id,
-                    request: MediaUploadRequestFfi(
-                        attachments: mediaAttachments.map(\.uploadRequest),
-                        caption: text.isEmpty ? nil : text,
-                        send: true,
-                        blossomServer: nil
-                    )
-                )
-            } else if let replyDraftContext {
-                _ = try await client.replyToMessage(
-                    accountRef: activeAccount.accountRef,
-                    groupIdHex: selectedChat.id,
-                    targetMessageId: replyDraftContext.targetMessageId,
-                    text: text
-                )
-            } else {
-                _ = try await client.sendText(
-                    accountRef: activeAccount.accountRef,
-                    groupIdHex: selectedChat.id,
-                    text: text
-                )
-            }
-            draftText = ""
-            replyDraftContext = nil
-            pendingMediaAttachmentsByConversation[draftKey] = nil
-            await refreshSelectedTimelineAfterSend(groupIdHex: selectedChat.id, account: activeAccount, client: client)
-        } catch {
-            lastError = error.localizedDescription
-        }
-    }
-
     func restoreOrSelectFirstAccount() {
         if let activeAccountId, accounts.contains(where: { $0.id == activeAccountId }) {
             return
@@ -2120,213 +1821,6 @@ final class WorkspaceState {
         privacySecuritySettings.auditLogCredentialsAvailable = config.auditLogCredentialsAvailable
     }
 
-    func startTimelineListener(
-        groupIdHex: String,
-        account: AccountItem,
-        subscription: TimelineMessagesSubscription? = nil
-    ) {
-        guard client != nil else { return }
-        stopTimelineListener()
-        guard activeAccountId == account.id, selectedChat?.id == groupIdHex else { return }
-        timelineTaskGroupId = groupIdHex
-        timelineTask = Task { [weak self] in
-            await self?.runTimelineListener(
-                groupIdHex: groupIdHex,
-                account: account,
-                existingSubscription: subscription
-            )
-        }
-    }
-
-    func stopTimelineListener() {
-        timelineTask?.cancel()
-        timelineTask = nil
-        timelineTaskGroupId = nil
-        activeTimelineSubscription = nil
-        activeTimelineGroupId = nil
-    }
-
-    func runTimelineListener(
-        groupIdHex: String,
-        account: AccountItem,
-        existingSubscription: TimelineMessagesSubscription? = nil
-    ) async {
-        guard let client else { return }
-        var reconnectAttempt = 0
-        var pendingSubscription = existingSubscription
-
-        while !Task.isCancelled,
-            activeAccountId == account.id,
-            selectedChat?.id == groupIdHex
-        {
-            do {
-                let subscription: TimelineMessagesSubscription
-                if let existing = pendingSubscription {
-                    subscription = existing
-                    pendingSubscription = nil
-                } else {
-                    subscription = try await client.subscribeTimelineMessages(
-                        accountRef: account.accountRef,
-                        groupIdHex: groupIdHex,
-                        limit: Self.timelinePageLimit
-                    )
-                    guard activeAccountId == account.id,
-                        selectedChat?.id == groupIdHex,
-                        !Task.isCancelled
-                    else { break }
-                    activeTimelineSubscription = subscription
-                    activeTimelineGroupId = groupIdHex
-                    if let page = subscription.snapshot() {
-                        await applyTimelineWindow(
-                            page,
-                            groupIdHex: groupIdHex,
-                            account: account,
-                            client: client
-                        )
-                    }
-                }
-                // `next()` blocks for the next live change and returns the resulting
-                // authoritative window (ordering, dedup, head-anchoring while scrolled back,
-                // and the cap are all owned by the runtime), so we render it directly.
-                while !Task.isCancelled,
-                    activeAccountId == account.id,
-                    selectedChat?.id == groupIdHex
-                {
-                    guard let page = await subscription.next() else { break }
-                    guard !Task.isCancelled,
-                        activeAccountId == account.id,
-                        selectedChat?.id == groupIdHex
-                    else { break }
-                    reconnectAttempt = 0
-                    await applyTimelineWindow(
-                        page,
-                        groupIdHex: groupIdHex,
-                        account: account,
-                        client: client
-                    )
-                }
-            } catch is CancellationError {
-                return
-            } catch {
-                if activeAccountId == account.id, selectedChat?.id == groupIdHex {
-                    setBackgroundStatus(error.localizedDescription)
-                }
-            }
-
-            guard !Task.isCancelled,
-                activeAccountId == account.id,
-                selectedChat?.id == groupIdHex
-            else { break }
-            do {
-                try await waitBeforeListenerReconnect(attempt: reconnectAttempt)
-            } catch is CancellationError {
-                return
-            } catch {
-                if activeAccountId == account.id, selectedChat?.id == groupIdHex {
-                    setBackgroundStatus(error.localizedDescription)
-                }
-            }
-            reconnectAttempt += 1
-        }
-
-        if timelineTaskGroupId == groupIdHex && !Task.isCancelled {
-            timelineTask = nil
-            timelineTaskGroupId = nil
-        }
-    }
-
-    func replaceMessages(
-        _ messages: [MessageItem],
-        groupIdHex: String,
-        paging: TimelinePagingState? = nil
-    ) {
-        // The window is already ordered, deduped, and capped by the runtime subscription,
-        // so render it as-is.
-        let nextPaging = paging ?? timelinePagingByChat[groupIdHex] ?? .empty
-        let messageLookup = Dictionary(uniqueKeysWithValues: messages.map { ($0.id, $0) })
-        let messageIDs = messages.map(\.id)
-
-        if messagesByChat.count == 1, messagesByChat[groupIdHex] != nil {
-            messagesByChat[groupIdHex] = messages
-            messageLookupByChat[groupIdHex] = messageLookup
-            messageIDsByChat[groupIdHex] = messageIDs
-        } else {
-            messagesByChat = [groupIdHex: messages]
-            messageLookupByChat = [groupIdHex: messageLookup]
-            messageIDsByChat = [groupIdHex: messageIDs]
-        }
-        if timelinePagingByChat.count == 1, timelinePagingByChat[groupIdHex] != nil {
-            timelinePagingByChat[groupIdHex] = nextPaging
-        } else {
-            timelinePagingByChat = [groupIdHex: nextPaging]
-        }
-        finishTimelineInitialLoad(groupIdHex: groupIdHex)
-    }
-
-    func refreshSelectedTimelineAfterSend(
-        groupIdHex: String,
-        account: AccountItem,
-        client: any MarmotRuntime
-    ) async {
-        guard activeAccountId == account.id, selectedChat?.id == groupIdHex else { return }
-        let pageLimit = Self.timelinePageLimit
-        do {
-            let page = try await runOffMain {
-                try client.timelineMessages(
-                    accountRef: account.accountRef,
-                    query: TimelineMessageQueryFfi(
-                        groupIdHex: groupIdHex,
-                        search: nil,
-                        before: nil,
-                        beforeMessageId: nil,
-                        after: nil,
-                        afterMessageId: nil,
-                        limit: pageLimit
-                    )
-                )
-            }
-            guard activeAccountId == account.id, selectedChat?.id == groupIdHex else { return }
-            await applyTimelineWindow(page, groupIdHex: groupIdHex, account: account, client: client)
-        } catch {
-            setBackgroundStatus(error.localizedDescription)
-        }
-    }
-
-    func pruneMessageCache(keeping groupIdHex: String?) {
-        defer {
-            pruneMediaDownloadCache(keeping: groupIdHex)
-        }
-
-        guard let groupIdHex else {
-            messagesByChat = [:]
-            messageLookupByChat = [:]
-            messageIDsByChat = [:]
-            timelinePagingByChat = [:]
-            timelineInitialLoadGroupId = nil
-            return
-        }
-
-        if let messages = messagesByChat[groupIdHex] {
-            messagesByChat = [groupIdHex: messages]
-            messageLookupByChat = [groupIdHex: Dictionary(uniqueKeysWithValues: messages.map { ($0.id, $0) })]
-            messageIDsByChat = [groupIdHex: messages.map(\.id)]
-        } else {
-            messagesByChat = [:]
-            messageLookupByChat = [:]
-            messageIDsByChat = [:]
-        }
-        if let paging = timelinePagingByChat[groupIdHex] {
-            timelinePagingByChat = [groupIdHex: paging]
-        } else {
-            timelinePagingByChat = [:]
-        }
-        if timelineInitialLoadGroupId != groupIdHex {
-            timelineInitialLoadGroupId = nil
-        } else if messagesByChat[groupIdHex] != nil {
-            timelineInitialLoadGroupId = nil
-        }
-    }
-
     func mediaDownloadKey(message: MessageItem, attachment: MessageMediaAttachment) -> String {
         [activeAccountId ?? "", message.groupIdHex, attachment.id].joined(separator: "\u{1F}")
     }
@@ -2371,96 +1865,6 @@ final class WorkspaceState {
             record.reference.plaintextSha256 == reference.plaintextSha256
                 || record.reference.ciphertextSha256 == reference.ciphertextSha256
         }?.reference ?? reference
-    }
-
-    func beginTimelineInitialLoadIfNeeded(groupIdHex: String) {
-        if messagesByChat[groupIdHex] == nil {
-            timelineInitialLoadGroupId = groupIdHex
-        } else if timelineInitialLoadGroupId == groupIdHex {
-            timelineInitialLoadGroupId = nil
-        }
-    }
-
-    func finishTimelineInitialLoad(groupIdHex: String) {
-        if timelineInitialLoadGroupId == groupIdHex {
-            timelineInitialLoadGroupId = nil
-        }
-    }
-
-    func markLatestVisibleMessageRead(
-        groupIdHex: String,
-        account: AccountItem,
-        client: any MarmotRuntime
-    ) async {
-        guard activeAccountId == account.id, selectedChat?.id == groupIdHex else { return }
-        // A selected chat is necessary but not sufficient: only advance the read marker
-        // when the app is active and the conversation window is actually visible. If the
-        // user has switched away or the window is hidden/minimized, incoming live deltas
-        // must not silently clear unread state for messages they have not seen — this
-        // mirrors the focus gate the notification path already applies in
-        // handleNotificationUpdate(_:). Marking is deferred until the conversation becomes
-        // visible again (see handleConversationVisibilityChange()).
-        guard selectedConversationIsVisible() else { return }
-        guard
-            let latest = (messagesByChat[groupIdHex] ?? []).last(where: { message in
-                message.timelineKind == 9 && !message.isDeleted
-            })
-        else {
-            return
-        }
-        let marker = ReadMarker(sentAt: latest.sentAt, messageId: latest.id)
-        let currentMarker = lastMarkedReadMarkers[groupIdHex]
-        guard currentMarker != marker else { return }
-        guard currentMarker.map({ $0 < marker }) ?? true else { return }
-        lastMarkedReadMarkers[groupIdHex] = marker
-
-        do {
-            let accountRef = account.accountRef
-            let messageId = latest.id
-            let row = try await runOffMain({
-                try client.markTimelineMessageRead(
-                    accountRef: accountRef,
-                    groupIdHex: groupIdHex,
-                    messageIdHex: messageId
-                )
-            })
-            guard activeAccountId == account.id, selectedChat?.id == groupIdHex else { return }
-            let committedState = ReadMarker.afterSuccessfulCommit(
-                current: lastMarkedReadMarkers[groupIdHex],
-                confirmed: lastConfirmedReadMarkers[groupIdHex],
-                attempted: marker
-            )
-            lastMarkedReadMarkers[groupIdHex] = committedState.current
-            lastConfirmedReadMarkers[groupIdHex] = committedState.confirmed
-            if let row {
-                await applyChatRow(row, account: account, shouldEnrich: false)
-            }
-        } catch {
-            guard activeAccountId == account.id, selectedChat?.id == groupIdHex else { return }
-            lastMarkedReadMarkers[groupIdHex] = ReadMarker.afterFailedOptimisticAdvance(
-                current: lastMarkedReadMarkers[groupIdHex],
-                attempted: marker,
-                confirmed: lastConfirmedReadMarkers[groupIdHex]
-            )
-            setBackgroundStatus(error.localizedDescription)
-        }
-    }
-
-    /// Flush any read-marking that was deferred while the conversation was not visible.
-    ///
-    /// `markLatestVisibleMessageRead(_:)` refuses to advance the read marker while the app
-    /// is inactive or its conversation window has no visible key window, so messages that
-    /// arrive while the user is away stay unread. When the conversation becomes visible
-    /// again it is safe to advance the marker to the latest visible message. Call this from
-    /// app/window activation hooks (see ContentView).
-    func handleConversationVisibilityChange() async {
-        guard selectedConversationIsVisible() else { return }
-        guard let client, let activeAccount, let selectedChat else { return }
-        await markLatestVisibleMessageRead(
-            groupIdHex: selectedChat.id,
-            account: activeAccount,
-            client: client
-        )
     }
 
     func resetNewChatComposer() {
